Added indexing for compilers.
Created data directory. Closed caption files are in data/compilers_caption. Generated index is written to data/auto_index.txt. The indexing code also uses the Dragon book index and a list of stop words, which live in the data directory.

The indexing code is in src/generate_index.ipynb.